% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplus.R
\docType{package}
\name{heatplus-package}
\alias{heatplus}
\alias{heatplus-package}
\title{HEAT Plus}
\description{
The HEAT Plus application.
}
\author{
\strong{Maintainer}: Zev Ross \email{zev@zevross.com}

Authors:
\itemize{
  \item Ahmadreza Hosseinpoor [contributor]
  \item Anne Schlotheuber [contributor]
}

Other contributors:
\itemize{
  \item Nate Teetor \email{nate@zevross.com} [contributor]
  \item  World Health Organization [copyright holder]
}

}
